refactor(a2a): merge startGoFixture into interop_launchers_test.go
The separate native_go_launchers_test.go file no longer serves a purpose
now that the orchestration layer is gone. Move startGoFixture alongside the
other startXYZFixture functions and delete the file.
